Embedding nidata into key_icon_struct.
